// one-cms/src/api/openapi/models/Space.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Server API - Login
 * The Restful APIs of Login.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface Space
 */
export interface Space {
  /**
   *
   * @type {string}
   * @memberof Space
   */
  spaceId?: string;
  /**
   *
   * @type {string}
   * @memberof Space
   */
  spaceGroupId?: string;
  /**
   *
   * @type {string}
   * @memberof Space
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof Space
   */
  description?: string;
  /**
   *
   * @type {string}
   * @memberof Space
   */
  thumbnail?: string;
  /**
   *
   * @type {string}
   * @memberof Space
   */
  addressable?: string;
  /**
   *
   * @type {Date}
   * @memberof Space
   */
  startAt?: Date;
  /**
   *
   * @type {Date}
   * @memberof Space
   */
  endAt?: Date;
}

/**
 * Check if a given object implements the Space interface.
 */
export function instanceOfSpace(value: object): boolean {
  let isInstance = true;

  return isInstance;
}

export function SpaceFromJSON(json: any): Space {
  return SpaceFromJSONTyped(json, false);
}

export function SpaceFromJSONTyped(json: any, ignoreDiscriminator: boolean): Space {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    spaceId: !exists(json, 'space_id') ? undefined : json['space_id'],
    spaceGroupId: !exists(json, 'space_group_id') ? undefined : json['space_group_id'],
    name: !exists(json, 'name') ? undefined : json['name'],
    description: !exists(json, 'description') ? undefined : json['description'],
    thumbnail: !exists(json, 'thumbnail') ? undefined : json['thumbnail'],
    addressable: !exists(json, 'addressable') ? undefined : json['addressable'],
    startAt: !exists(json, 'start_at') ? undefined : new Date(json['start_at']),
    endAt: !exists(json, 'end_at') ? undefined : new Date(json['end_at']),
  };
}

export function SpaceToJSON(value?: Space | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    space_id: value.spaceId,
    space_group_id: value.spaceGroupId,
    name: value.name,
    description: value.description,
    thumbnail: value.thumbnail,
    addressable: value.addressable,
    start_at: value.startAt === undefined ? undefined : value.startAt.toISOString(),
    end_at: value.endAt === undefined ? undefined : value.endAt.toISOString(),
  };
}